Pass separator character to TagLib functions
This is a bit of a hack and assumes my traditional semi-colon separator
(rather than comma or slash), but it feels like a real implementation of
this requires significant effort in both TagLib and Gerbera.  This proof
of concept should fit my needs and provide a starting point for further
discussions if needed.